Platillos (#1)
* upgrade spring boot

* integrate security with auth0

* use lombok, and create platillo and ingrediente entities

* draft templates for platillos

* passing initial Alimento controller unit test

* Tests completed for Alimento REST controller

* Test completed for AlimentosController

* PlatilloRestController draft tested

* draft ingredientes

* basic functionality to add ingredients

* agregar ingrediente y refactor de controlador de dietas

* add and remove ingredientes

* upload platillo picture

* upload and display platillo's pictures

* pdf, video and ingesta buttons

* modal para agregar ingesta de platillo

* pdf upload modal

* video modal

* ingestas recomendadas

* add preparation instructions in PDF and video

* Try to setup Java 17

* default values for test classes

+++ src/main/java/com/nutriconsultas/consulta/ConsultaRestController.java
@@ -8,6 +8,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import io.micrometer.common.lang.NonNull;
+
 @RestController
 @RequestMapping("/rest")
 public class ConsultaRestController {
@@ -17,9 +19,13 @@ public class ConsultaRestController {
   private ConsultaRepository repo;
 
   @DeleteMapping("consultas/{id}")
-  public String deleteConsulta(@PathVariable("id") Long id) {
-    repo.deleteById(id);
-    logger.info("La consulta {} ha sido borrada", id);
-    return "OK";
+  public String deleteConsulta(@PathVariable @NonNull Long id) {
+    if (id != null) {
+      repo.deleteById(id);
+      logger.info("La consulta {} ha sido borrada", id);
+      return "OK";
+    } else {
+      return "Invalid ID";
+    }
   }
 }
